new/terraform_v2/tf_composer_code/DestroyCloudFunction/requirements.txt
google-cloud-storage
cryptography==2.8
crcmod
google-cloud-secret-manager